Move security-related classes from Hosting to Security
